My bad, reverting.

+++ trunk/src/helper/Arithmetic.java
@@ -7,7 +7,7 @@ package helper;
  */
 public class Arithmetic {
 	
-	public static long factorial(int n){
+	public static double factorial(double n){
 		if(n==0||n==1) return 1;
 		return n*(factorial(n-1));
 	}
